Finished draft of invitations.

+++ Database/Schema/Technician/Tables/Positions.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE Technician.Positions
 (
-	Id TINYINT NOT NULL IDENTITY (0, 1),
+	Id INT NOT NULL IDENTITY (0, 1),
 	[Name] NVARCHAR(50) NOT NULL,
 	BusinessId INT NOT NULL,
 	CONSTRAINT [PK_Technician_Positions] PRIMARY KEY CLUSTERED (Id),
